Move existing methods into class
